Refactored gothic window code and implemented some style examples
Refactored gothic window function to receive a structure instead of specific values. This structure will be the basis to attach a given style to a function, which will, in turn, allow for other fields to incorporate additional styles (i.e., rosettes, sub-arches, among others). Use cases have been implemented, together with a migration to the cathedral file to later be applied to cathedral walls/windows